backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
pillow>=10.2.0
numpy>=1.26.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
torch>=2.2.0
torchvision>=0.17.0
